api_test/requirements.txt
google-generativeai
pillow
python-dotenv
pytesseract
replicate
